+++ contracts/Clans/ClanBattleLibrary.sol
@@ -11,9 +11,9 @@ import {IPlayers} from "../interfaces/IPlayers.sol";
 import "../globals/all.sol";
 
 library ClanBattleLibrary {
-  function shuffleArray(uint48[] memory _array, uint _randomNumber) public pure returns (uint48[] memory) {
-    for (uint i; i < _array.length; ++i) {
-      uint n = i + (_randomNumber % (_array.length - i));
+  function shuffleArray(uint48[] memory _array, uint256 _randomNumber) public pure returns (uint48[] memory) {
+    for (uint256 i; i < _array.length; ++i) {
+      uint256 n = i + (_randomNumber % (_array.length - i));
       if (i != n) {
         uint48 temp = _array[n];
         _array[n] = _array[i];
@@ -28,10 +28,10 @@ library ClanBattleLibrary {
     uint48[] memory _clanMembersA,
     uint48[] memory _clanMembersB,
     uint8[] memory _skills,
-    uint _randomWordA,
-    uint _randomWordB,
-    uint _extraRollsA,
-    uint _extraRollsB
+    uint256 _randomWordA,
+    uint256 _randomWordB,
+    uint256 _extraRollsA,
+    uint256 _extraRollsB
   )
     external
     view
@@ -43,13 +43,13 @@ library ClanBattleLibrary {
     )
   {
     Skill[] memory skills = new Skill[](_skills.length);
-    for (uint i; i < _skills.length; ++i) {
+    for (uint256 i; i < _skills.length; ++i) {
       skills[i] = Skill(_skills[i]);
     }
 
     BattleResultEnum[] memory battleResultsEnum;
-    uint[] memory rollsA;
-    uint[] memory rollsB;
+    uint256[] memory rollsA;
+    uint256[] memory rollsB;
     (battleResultsEnum, rollsA, rollsB, didAWin) = doBattle(
       _players,
       _clanMembersA,
@@ -61,7 +61,7 @@ library ClanBattleLibrary {
     );
 
     battleResults = new uint8[](battleResultsEnum.length);
-    for (uint i; i < battleResultsEnum.length; ++i) {
+    for (uint256 i; i < battleResultsEnum.length; ++i) {
       battleResults[i] = uint8(battleResultsEnum[i]);
     }
 
@@ -74,54 +74,54 @@ library ClanBattleLibrary {
     uint48[] memory _clanMembersA, // [In/Out] gets shuffled
     uint48[] memory _clanMembersB, // [In/Out] gets shuffled
     Skill[] memory _skills,
-    uint[2] memory _randomWords,
-    uint _extraRollsA,
-    uint _extraRollsB
+    uint256[2] memory _randomWords,
+    uint256 _extraRollsA,
+    uint256 _extraRollsB
   )
     internal
     view
-    returns (BattleResultEnum[] memory battleResults, uint[] memory rollsA, uint[] memory rollsB, bool didAWin)
+    returns (BattleResultEnum[] memory battleResults, uint256[] memory rollsA, uint256[] memory rollsB, bool didAWin)
   {
     shuffleArray(_clanMembersA, _randomWords[0]);
     shuffleArray(_clanMembersB, _randomWords[1]);
 
-    uint baseClanMembersCount = _clanMembersA.length > _clanMembersB.length
+    uint256 baseClanMembersCount = _clanMembersA.length > _clanMembersB.length
       ? _clanMembersB.length
       : _clanMembersA.length;
 
     battleResults = new BattleResultEnum[](Math.max(_clanMembersA.length, _clanMembersB.length));
-    rollsA = new uint[](Math.max(_clanMembersA.length, _clanMembersB.length));
-    rollsB = new uint[](Math.max(_clanMembersA.length, _clanMembersB.length));
+    rollsA = new uint256[](Math.max(_clanMembersA.length, _clanMembersB.length));
+    rollsB = new uint256[](Math.max(_clanMembersA.length, _clanMembersB.length));
 
-    uint numWinnersA;
-    uint numWinnersB;
+    uint256 numWinnersA;
+    uint256 numWinnersB;
     {
-      for (uint i; i < baseClanMembersCount; ++i) {
+      for (uint256 i; i < baseClanMembersCount; ++i) {
         // It's possible that there are empty entries if they left the clan
         if (_clanMembersA[i] == 0 || _clanMembersB[i] == 0) {
           rollsA[i] = _clanMembersA[i] == 0 ? 0 : 1;
           rollsB[i] = _clanMembersB[i] == 0 ? 0 : 1;
         } else {
           {
-            uint levelA = PlayersLibrary._getLevel(IPlayers(_players).xp(_clanMembersA[i], _skills[i]));
-            uint levelB = PlayersLibrary._getLevel(IPlayers(_players).xp(_clanMembersB[i], _skills[i]));
+            uint256 levelA = PlayersLibrary._getLevel(IPlayers(_players).xp(_clanMembersA[i], _skills[i]));
+            uint256 levelB = PlayersLibrary._getLevel(IPlayers(_players).xp(_clanMembersB[i], _skills[i]));
             if (levelA > 20 * 6 || levelB > 20 * 6) {
               assert(false); // Unsupported
             }
 
-            uint numRollsA = (levelA / 20) +
+            uint256 numRollsA = (levelA / 20) +
               (IPlayers(_players).isPlayerUpgraded(_clanMembersA[i]) ? 2 : 1) +
               _extraRollsA;
-            uint numRollsB = (levelB / 20) +
+            uint256 numRollsB = (levelB / 20) +
               (IPlayers(_players).isPlayerUpgraded(_clanMembersB[i]) ? 2 : 1) +
               _extraRollsB;
             bytes1 byteA = bytes32(_randomWords[0])[31 - i];
             // Check how many bits are set based on the number of rolls
-            for (uint j; j < numRollsA; ++j) {
+            for (uint256 j; j < numRollsA; ++j) {
               rollsA[i] += uint8(byteA >> j) & 1;
             }
             bytes1 byteB = bytes32(_randomWords[1])[31 - i];
-            for (uint j; j < numRollsB; ++j) {
+            for (uint256 j; j < numRollsB; ++j) {
               rollsB[i] += uint8(byteB >> j) & 1;
             }
           }
@@ -140,12 +140,12 @@ library ClanBattleLibrary {
 
     if (_clanMembersB.length > _clanMembersA.length) {
       numWinnersB += _clanMembersB.length - _clanMembersA.length;
-      for (uint i = baseClanMembersCount; i < _clanMembersB.length; ++i) {
+      for (uint256 i = baseClanMembersCount; i < _clanMembersB.length; ++i) {
         battleResults[i] = BattleResultEnum.LOSE;
       }
     } else if (_clanMembersA.length > _clanMembersB.length) {
       numWinnersA += _clanMembersA.length - _clanMembersB.length;
-      for (uint i = baseClanMembersCount; i < _clanMembersA.length; ++i) {
+      for (uint256 i = baseClanMembersCount; i < _clanMembersA.length; ++i) {
         battleResults[i] = BattleResultEnum.WIN;
       }
     }
