+++ contracts/EstforLibrary.sol
@@ -11,7 +11,7 @@ import "./globals/all.sol";
 // This file contains methods for interacting with generic functions like trimming strings, lowercase etc.
 library EstforLibrary {
   using UnsafeMath for U256;
-  using UnsafeMath for uint;
+  using UnsafeMath for uint256;
 
   function isWhitespace(bytes1 _char) internal pure returns (bool) {
     return
@@ -25,10 +25,10 @@ library EstforLibrary {
 
   function leftTrim(string memory _str) internal pure returns (string memory) {
     bytes memory b = bytes(_str);
-    uint strLen = b.length;
-    uint start = type(uint).max;
+    uint256 strLen = b.length;
+    uint256 start = type(uint256).max;
     // Find the index of the first non-whitespace character
-    for (uint i = 0; i < strLen; ++i) {
+    for (uint256 i = 0; i < strLen; ++i) {
       bytes1 char = b[i];
       if (!isWhitespace(char)) {
         start = i;
@@ -36,12 +36,12 @@ library EstforLibrary {
       }
     }
 
-    if (start == type(uint).max) {
+    if (start == type(uint256).max) {
       return "";
     }
     // Copy the remainder to a new string
     bytes memory trimmedBytes = new bytes(strLen - start);
-    for (uint i = start; i < strLen; ++i) {
+    for (uint256 i = start; i < strLen; ++i) {
       trimmedBytes[i - start] = b[i];
     }
     return string(trimmedBytes);
@@ -49,14 +49,14 @@ library EstforLibrary {
 
   function rightTrim(string calldata _str) internal pure returns (string memory) {
     bytes memory b = bytes(_str);
-    uint strLen = b.length;
+    uint256 strLen = b.length;
     if (strLen == 0) {
       return "";
     }
     int end = -1;
     // Find the index of the last non-whitespace character
     for (int i = int(strLen) - 1; i >= 0; --i) {
-      bytes1 char = b[uint(i)];
+      bytes1 char = b[uint256(i)];
       if (!isWhitespace(char)) {
         end = i;
         break;
@@ -67,8 +67,8 @@ library EstforLibrary {
       return "";
     }
 
-    bytes memory trimmedBytes = new bytes(uint(end) + 1);
-    for (uint i = 0; i <= uint(end); ++i) {
+    bytes memory trimmedBytes = new bytes(uint256(end) + 1);
+    for (uint256 i = 0; i <= uint256(end); ++i) {
       trimmedBytes[i] = b[i];
     }
     return string(trimmedBytes);
@@ -85,7 +85,7 @@ library EstforLibrary {
     U256 iter = b.length.asU256();
     while (iter.neq(0)) {
       iter = iter.dec();
-      uint i = iter.asUint256();
+      uint256 i = iter.asUint256();
       bytes1 char = b[i];
 
       bool isUpperCaseLetter = (char >= 0x41) && (char <= 0x5A); // A-Z
@@ -107,7 +107,7 @@ library EstforLibrary {
     U256 iter = discord.length.asU256();
     while (iter.neq(0)) {
       iter = iter.dec();
-      uint i = iter.asUint256();
+      uint256 i = iter.asUint256();
       bytes1 char = discord[i];
 
       bool isUpperCaseLetter = (char >= 0x41) && (char <= 0x5A); // A-Z
@@ -126,7 +126,7 @@ library EstforLibrary {
     U256 iter = telegram.length.asU256();
     while (iter.neq(0)) {
       iter = iter.dec();
-      uint i = iter.asUint256();
+      uint256 i = iter.asUint256();
       bytes1 char = telegram[i];
 
       bool isUpperCaseLetter = (char >= 0x41) && (char <= 0x5A); // A-Z
@@ -146,7 +146,7 @@ library EstforLibrary {
     U256 iter = twitter.length.asU256();
     while (iter.neq(0)) {
       iter = iter.dec();
-      uint i = iter.asUint256();
+      uint256 i = iter.asUint256();
       bytes1 char = twitter[i];
 
       bool isUpperCaseLetter = (char >= 0x41) && (char <= 0x5A); // A-Z
@@ -165,7 +165,7 @@ library EstforLibrary {
     U256 iter = socialMediaName.length.asU256();
     while (iter.neq(0)) {
       iter = iter.dec();
-      uint i = iter.asUint256();
+      uint256 i = iter.asUint256();
       bytes1 char = socialMediaName[i];
 
       bool isUpperCaseLetter = (char >= 0x41) && (char <= 0x5A); // A-Z
@@ -187,7 +187,7 @@ library EstforLibrary {
     U256 iter = lowerStr.length.asU256();
     while (iter.neq(0)) {
       iter = iter.dec();
-      uint i = iter.asUint256();
+      uint256 i = iter.asUint256();
       if ((uint8(lowerStr[i]) >= 65) && (uint8(lowerStr[i]) <= 90)) {
         // So we add 32 to make it lowercase
         lowerStr[i] = bytes1(uint8(lowerStr[i]) + 32);
@@ -197,12 +197,12 @@ library EstforLibrary {
   }
 
   // This should match the one below, useful when a calldata array is needed and for external testing
-  function binarySearchMemory(uint48[] calldata _arr, uint _target) external pure returns (uint) {
-    uint low = 0;
-    uint high = _arr.length - 1;
+  function binarySearchMemory(uint48[] calldata _arr, uint256 _target) external pure returns (uint256) {
+    uint256 low = 0;
+    uint256 high = _arr.length - 1;
 
     while (low <= high) {
-      uint mid = low + (high - low) / 2;
+      uint256 mid = low + (high - low) / 2;
 
       if (_arr[mid] == _target) {
         return mid; // Element found
@@ -219,16 +219,16 @@ library EstforLibrary {
       }
     }
 
-    return type(uint).max; // Element not found
+    return type(uint256).max; // Element not found
   }
 
   // This should match the one above
-  function _binarySearch(uint48[] storage _arr, uint _target) internal view returns (uint) {
-    uint low = 0;
-    uint high = _arr.length - 1;
+  function _binarySearch(uint48[] storage _arr, uint256 _target) internal view returns (uint256) {
+    uint256 low = 0;
+    uint256 high = _arr.length - 1;
 
     while (low <= high) {
-      uint mid = low + (high - low) / 2;
+      uint256 mid = low + (high - low) / 2;
 
       if (_arr[mid] == _target) {
         return mid; // Element found
@@ -245,10 +245,10 @@ library EstforLibrary {
       }
     }
 
-    return type(uint).max; // Element not found
+    return type(uint256).max; // Element not found
   }
 
-  function binarySearch(uint48[] storage _arr, uint _target) external view returns (uint) {
+  function binarySearch(uint48[] storage _arr, uint256 _target) external view returns (uint256) {
     return _binarySearch(_arr, _target);
   }
 }
